Changes made as discribed
the main tag on lines 13/17  and footer tag 19/21 should be indented one tab to be in line with the header tag?
I was thinking all of these are element and have same priority  that's why should be indented in same vertical line.  Also there is no parent chield relationship here.